backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
billiard==4.2.4
cachetools==6.2.6
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.5
cron-descriptor==1.4.5
cryptography==46.0.5
deprecation==2.1.0
dj-database-url==3.1.2
Django==4.2.29
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
fsspec==2026.2.0
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
postgrest==2.28.3
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
pytest==9.0.2
pytest-cov==7.1.0
pytest-django==4.12.0
pytest-mock==3.15.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
realtime==2.28.3
redis==7.4.0
requests==2.32.5
rich==14.3.3
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.5
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
stripe==14.4.1
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.7
whitenoise==6.12.0
yarl==1.23.0
zstandard==0.25.0
django-storages==1.15.0
boto3==1.36.26
